skeleton tweaks
