+++ src/assets/labels.js
@@ -3,6 +3,7 @@ export default {
     logo: "./src/assets/img/atomiql-logo.png",
     headTitle: "High performance meets scalable flexibility",
     headSubtitle: "GraphQL client designed for Jotai",
+    btnStart: "Get Started",
   },
   about:
     "AtomiQL is the first GraphQL client that allows developers to reap the benefits of type-safety, UI and data layer logic separation, and a clean query language interface, while also maximizing front-end rendering performance with atomic state management. Especially with this increasing nature of frontend complexity, shifting towards atomic state management with the Jotai library can significantly reduce page load time and improve user experience through eliminating unnecessary re-renders.",
