+++ src/containers/About.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import labels from './../assets/labels.js';
+
+class About extends React.Component {
+    render() {
+        return (
+            <div id="about">
+                {labels.about}
+            </div>
+        )
+    }
+}
+
+export default About;
