+++ src/containers/Demo.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import labels from './../assets/labels.js';
+
+class Demo extends React.Component {
+    render() {
+        return (
+            <div id="demo">
+                demo placeholder
+            </div>
+        )
+    }
+}
+
+export default Demo;
